Ray ready to merge
