+++ p2pfl/learning/__init__.py
@@ -10,6 +10,7 @@ import importlib
 from p2pfl.learning.dataset.p2pfl_dataset import P2PFLDataset
 from p2pfl.learning.learner import NodeLearner
 from p2pfl.learning.p2pfl_model import P2PFLModel
+from p2pfl.utils.check_ray import ray_installed
 
 
 def try_init_learner_with_ray(learner: type[NodeLearner], model: P2PFLModel, data: P2PFLDataset, addr: str):
@@ -27,7 +28,7 @@ def try_init_learner_with_ray(learner: type[NodeLearner], model: P2PFLModel, dat
 
     """
     learner_instance = learner(model, data, addr)
-    if importlib.util.find_spec("ray") is not None:
+    if ray_installed():
         from p2pfl.learning.simulation.virtual_learner import VirtualNodeLearner
 
         learner_instance = VirtualNodeLearner(learner_instance, addr)
