+++ p2pfl/learning/simulation/virtual_learner.py
@@ -17,7 +17,6 @@
 #
 """Virtual Node Learner."""
 
-import contextlib
 from typing import Dict, List, Union
 
 import numpy as np
@@ -115,14 +114,8 @@ class VirtualNodeLearner(NodeLearner):
 
     def interrupt_fit(self) -> None:
         """Interrupt the fit process."""
-        self.actor_pool.interrupt_submit(str(self.addr))
-        """
-        with contextlib.suppress(KeyError): # If training is not running
-            .submit(
-                lambda actor, addr, learner: actor.interrupt_fit.remote(addr, learner),
-                (str(self.addr), self.learner),
-            )
-        """
+        # TODO: Need to implement this!
+        raise NotImplementedError
 
     def evaluate(self) -> Dict[str, float]:
         """
