+++ p2pfl/management/logger/decorators/async_logger.py
@@ -22,35 +22,37 @@ import logging
 import multiprocessing
 from logging.handlers import QueueHandler, QueueListener
 
+from p2pfl.management.logger.decorators.logger_decorator import LoggerDecorator
 from p2pfl.management.logger.logger import P2PFLogger
-from p2pfl.management.logger.loggers.logger_decorator import P2PFLoggerDecorator
 
 
-class AsyncLocalLogger(P2PFLoggerDecorator):
+class AsyncLogger(LoggerDecorator):
     """Async logger decorator."""
-
-    _p2pflogger: P2PFLogger
 
     def __init__(self, p2pflogger: P2PFLogger) -> None:
         """Initialize the logger."""
-        self._p2pflogger = p2pflogger
+        super().__init__(p2pflogger)
 
         # Set up asynchronous logging
         self.log_queue: multiprocessing.Queue[logging.LogRecord] = multiprocessing.Queue()
         queue_handler = QueueHandler(self.log_queue)
-        self.get_logger().addHandler(queue_handler)
+        self._p2pfl_logger.add_handler(queue_handler)
 
         # Set up a listener for the queue
-        self.queue_listener = QueueListener(self.log_queue, *self.get_handlers())
+        self.queue_listener = QueueListener(self.log_queue)
         self.queue_listener.start()
 
         # Register cleanup function to close the queue on exit
         atexit.register(self.cleanup)
+
+    def add_handler(self, handler):
+        """Add a handler to the logger."""
+        # add handler to tuple of handlers
+        self.queue_listener.handlers = self.queue_listener.handlers + (handler,)
 
     def cleanup(self) -> None:
         """Cleanup the logger."""
         if self.queue_listener:
             self.queue_listener.stop()
         self.log_queue.close()
-
-        self._p2pflogger.cleanup()
+        super().cleanup()
