+++ p2pfl/management/logger/decorators/logger_decorator.py
@@ -1,0 +1,243 @@
+#
+# This file is part of the federated_learning_p2p (p2pfl) distribution
+# (see https://github.com/pguijas/federated_learning_p2p).
+# Copyright (c) 2022 Pedro Guijas Bravo.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, version 3.
+#
+# This program is distributed in the hope that it will be useful, but
+# WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+# General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
+#
+
+"""Decorator for the logger to be used in the simulation."""
+
+import datetime
+import logging
+from typing import Any, Dict, Union
+
+from p2pfl.experiment import Experiment
+from p2pfl.management.logger.logger import P2PFLogger
+from p2pfl.management.metric_storage import GlobalLogsType, LocalLogsType
+
+
+class LoggerDecorator(P2PFLogger):
+    """
+    Decorator class for logging. Works by wrapping the a previous logger, so new funcionalities can be added.
+
+    By default, it does nothing, just delegates the calls to the wrapped logger.
+    """
+
+    def __init__(self, logger: P2PFLogger) -> None:
+        """
+        Initialize the logger.
+
+        Args:
+            logger: The logger to wrap.
+
+        """
+        self._p2pfl_logger = logger
+
+    def connect_web(self, url: str, key: str) -> None:
+        """
+        Connect to the web services.
+
+        Args:
+            url: The URL of the web services.
+            key: The API key.
+
+        """
+        self._p2pfl_logger.connect_web(url, key)
+
+    def cleanup(self) -> None:
+        """Cleanup the logger."""
+        self._p2pfl_logger.cleanup()
+
+    def set_level(self, level: Union[int, str]) -> None:
+        """
+        Set the logger level.
+
+        Args:
+            level: The logger level.
+
+        """
+        self._p2pfl_logger.set_level(level)
+
+    def get_level(self) -> int:
+        """
+        Get the logger level.
+
+        Returns
+            The logger level.
+
+        """
+        return self._p2pfl_logger.get_level()
+
+    def get_level_name(self, lvl: int) -> str:
+        """
+        Get the logger level name.
+
+        Args:
+            lvl: The logger level.
+
+        Returns:
+            The logger level name.
+
+        """
+        return self._p2pfl_logger.get_level_name(lvl)
+
+    def log(self, level: int, node: str, message: str) -> None:
+        """
+        Log a message.
+
+        Args:
+            level: The log level.
+            node: The node name.
+            message: The message to log.
+
+        """
+        self._p2pfl_logger.log(level, node, message)
+
+    def log_metric(self, addr: str, metric: str, value: float, round: int | None = None, step: int | None = None) -> None:
+        """
+        Log a metric.
+
+        Args:
+            addr: The node name.
+            metric: The metric to log.
+            value: The value.
+            step: The step.
+            round: The round.
+
+        """
+        self._p2pfl_logger.log_metric(addr, metric, value, round, step)
+
+    def get_local_logs(self) -> LocalLogsType:
+        """
+        Get the logs.
+
+        Args:
+            node: The node name.
+            exp: The experiment name.
+
+        Returns:
+            The logs.
+
+        """
+        return self._p2pfl_logger.get_local_logs()
+
+    def get_global_logs(self) -> GlobalLogsType:
+        """
+        Get the logs.
+
+        Args:
+            node: The node name.
+            exp: The experiment name.
+
+        Returns:
+            The logs.
+
+        """
+        return self._p2pfl_logger.get_global_logs()
+
+    def register_node(self, node: str, simulation: bool) -> None:
+        """
+        Register a node.
+
+        Args:
+            node: The node address.
+            simulation: If the node is a simulation.
+
+        """
+        self._p2pfl_logger.register_node(node, simulation)
+
+    def unregister_node(self, node: str) -> None:
+        """
+        Unregister a node.
+
+        Args:
+            node: The node address.
+
+        """
+        self._p2pfl_logger.unregister_node(node)
+
+    def experiment_started(self, node: str, experiment: Experiment | None) -> None:
+        """
+        Notify the experiment start.
+
+        Args:
+            node: The node address.
+            experiment: The experiment.
+
+        """
+        self._p2pfl_logger.experiment_started(node, experiment)
+
+    def experiment_finished(self, node: str) -> None:
+        """
+        Notify the experiment end.
+
+        Args:
+            node: The node address.
+
+        """
+        self._p2pfl_logger.experiment_finished(node)
+
+    def round_started(self, node: str, experiment: Experiment | None) -> None:
+        """
+        Notify the round start.
+
+        Args:
+            node: The node address.
+            experiment: The experiment.
+
+        """
+        self._p2pfl_logger.round_started(node, experiment)
+
+    def round_finished(self, node: str) -> None:
+        """
+        Notify the round end.
+
+        Args:
+            node: The node address.
+
+        """
+        self._p2pfl_logger.round_finished(node)
+
+    def get_nodes(self) -> Dict[str, Dict[Any, Any]]:
+        """
+        Get the registered nodes.
+
+        Returns:
+            The registered nodes.
+
+        """
+        return self._p2pfl_logger.get_nodes()
+
+    def add_handler(self, handler: logging.Handler) -> None:
+        """
+        Add a handler to the logger.
+
+        Args:
+            handler: The handler to add.
+
+        """
+        self._p2pfl_logger.add_handler(handler)
+
+    def log_system_metric(self, node: str, metric: str, value: float, time: datetime.datetime) -> None:
+        """
+        Log a system metric.
+
+        Args:
+            node: The node name.
+            metric: The metric to log.
+            value: The value.
+            time: The time.
+
+        """
+        self._p2pfl_logger.log_system_metric(node, metric, value, time)
